Add a configuration to use the configured weapon chance when falling back
This deals with a problem if mods added armor for enemies that can't be legendary, for example, because otherwise the number of legendary weapons will be very high. Changing this setting would affect mod compatibility if anything modifies the base legendary list, though, which includes Power Armor to the People.